requirements.txt
# --- CORE DEEP LEARNING (VISION) ---
tensorflow==2.16.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0

# --- CORE DEEP LEARNING (LLM & NLP) ---
torch==2.3.0
transformers==4.41.2
accelerate==0.34.2
bitsandbytes==0.43.3
sentence-transformers==3.1.1
huggingface-hub==0.25.1

# --- RAG FRAMEWORK ---
langchain==0.3.0
langchain-huggingface==0.1.0
langchain-community==0.3.0
langchain-core==0.3.0
faiss-cpu==1.8.0

# --- MLOPS & TRACKING ---
mlflow==2.16.2

# --- UI & DEPLOYMENT ---
streamlit==1.38.0
python-dotenv==1.0.1

# --- TUNING HYPERPARAMETER ---
optuna